Restructured the UI package by feature
