+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
     android:roundIcon="@mipmap/ic_launcher_round"
     android:supportsRtl="true"
     android:theme="@style/AppTheme">
-    <activity android:name="com.livehappyapps.githubviewer.ui.activity.SettingsActivity" />
-    <activity android:name="com.livehappyapps.githubviewer.ui.activity.IssueActivity" />
+    <activity android:name="com.livehappyapps.githubviewer.ui.settings.SettingsActivity" />
+    <activity android:name="com.livehappyapps.githubviewer.ui.issues.IssueActivity" />
     <activity
-      android:name="com.livehappyapps.githubviewer.ui.activity.MainActivity"
+      android:name="com.livehappyapps.githubviewer.ui.main.MainActivity"
       android:theme="@style/AppTheme.NoActionBar">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
