+++ app/src/main/java/com/livehappyapps/githubviewer/ui/issues/IssueAdapter.kt
@@ -1,4 +1,4 @@
-package com.livehappyapps.githubviewer.ui.adapter
+package com.livehappyapps.githubviewer.ui.issues
 
 import android.graphics.Color
 import android.view.LayoutInflater
