+++ app/src/main/java/com/livehappyapps/githubviewer/ui/issues/IssuePageAdapter.kt
@@ -1,10 +1,9 @@
-package com.livehappyapps.githubviewer.ui.adapter
+package com.livehappyapps.githubviewer.ui.issues
 
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.livehappyapps.githubviewer.IssueState
 import com.livehappyapps.githubviewer.R
-import com.livehappyapps.githubviewer.ui.fragment.IssueFragment
 
 
 class IssuePagerAdapter(
