+++ app/src/main/java/com/livehappyapps/githubviewer/ui/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.livehappyapps.githubviewer.ui.viewmodel
+package com.livehappyapps.githubviewer.ui.main
 
 import android.content.SharedPreferences
 import androidx.lifecycle.LiveData
