+++ app/src/main/java/com/livehappyapps/githubviewer/ui/main/RepoAdapter.kt
@@ -1,4 +1,4 @@
-package com.livehappyapps.githubviewer.ui.adapter
+package com.livehappyapps.githubviewer.ui.main
 
 import android.text.Html
 import android.view.LayoutInflater
