+++ app/src/main/res/layout/activity_issue.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.activity.IssueActivity">
+  tools:context=".ui.issues.IssueActivity">
 
   <com.google.android.material.tabs.TabLayout
     android:id="@+id/tab_layout"
