+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.activity.MainActivity">
+  tools:context=".ui.main.MainActivity">
 
   <com.google.android.material.appbar.AppBarLayout
     android:layout_width="match_parent"
