+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.activity.SettingsActivity">
+  tools:context=".ui.settings.SettingsActivity">
 
   <FrameLayout
     android:id="@+id/container"
